X5_send_letter_UI_some_api_tests/requirements.txt
pytest==5.0.1
allure-pytest==2.8.5
selenium==3.141.0
requests==2.22.0

